+++ src/test/kotlin/com/vonage/client/kt/AbstractTest.kt
@@ -57,7 +57,12 @@ abstract class AbstractTest {
     protected val text = "Hello, World!"
     protected val country = "GB"
     protected val secret = "ABCDEFGH01234abc"
+    protected val cursor = "7EjDNQrAcipmOnc0HCzpQRkhBULzY44ljGUX4lXKyUIVfiZay5pv9wg="
+    protected val vbcExt = "4321"
+    protected val userName = "Sam_username"
     protected val sipUri = "sip:rebekka@sip.example.com"
+    protected val websocketUri = "wss://example.com/socket"
+    protected val wsContentType = "audio/l16;rate=8000"
     protected val clientRef = "my-personal-reference"
     protected val textHexEncoded = "48656c6c6f2c20576f726c6421"
     protected val entityId = "1101407360000017170"
@@ -84,6 +89,10 @@ abstract class AbstractTest {
     protected val statusCallbackUrl = "$callbackUrl/status"
     protected val moCallbackUrl = "$callbackUrl/inbound-sms"
     protected val drCallbackUrl = "$callbackUrl/delivery-receipt"
+    protected val imageUrl = "$exampleUrlBase/image.jpg"
+    protected val audioUrl = "$exampleUrlBase/audio.mp3"
+    protected val videoUrl = "$exampleUrlBase/video.mp4"
+    protected val fileUrl = "$exampleUrlBase/file.pdf"
 
     private val port = 8081
     private val wiremock: WireMockServer = WireMockServer(
@@ -281,7 +290,7 @@ abstract class AbstractTest {
 
     protected inline fun <reified E: VonageApiResponseException> assertApiResponseException(
             url: String, requestMethod: HttpMethod, actualCall: () -> Any, status: Int,
-            errorType: String? = null, title: String? = null,
+            errorType: String? = null, title: String? = null, code: String? = null,
             detail: String? = null, instance: String? = null): E {
 
         val responseParams = mutableMapOf<String, Any>()
@@ -289,6 +298,7 @@ abstract class AbstractTest {
         if (title != null) responseParams["title"] = title
         if (detail != null) responseParams["detail"] = detail
         if (instance != null) responseParams["instance"] = instance
+        if (code != null) responseParams["code"] = code
 
         mockRequest(requestMethod, url).mockReturn(status, responseParams)
         val exception = assertThrows<E> { actualCall.invoke() }
@@ -298,6 +308,7 @@ abstract class AbstractTest {
         assertEquals(title, exception.title)
         assertEquals(instance, exception.instance)
         assertEquals(detail, exception.detail)
+        assertEquals(code, exception.code)
         return exception
     }