+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [0.9.0] - 2024-08-2?
 
 ### Added
-- Application API
+- Application & Users API
 
 ## [0.8.0] - 2024-08-09